:sparkles: Allow to skip fetching secrets from AWS Secrets Manager
In GCP, we push secrets differently and we can't have this hard dependency. It would be ideal to have a config setting for that instead of ENV VAR but there are 2 reasons to do that:
* I don't want to update all existing projects using it
* It is not easy to read configuration value at this (early) stage. Possible, but it would require some more hacks to be done.

That's why it is not worth it and better to rely on ENV VAR in case of GCP deployments.