EZYC-4280 :sparkles: Allow connecting to identity server protected by invalid ssl cert
For now in GCP-Cert & GCP-Prod we can only rely on internal, sabre-gcp.com domains which in CERT & PROD will NOT get valid ssl.
This prevents the id-admin from starting as it fails with connecting to id-sts.
Adding this option will enable us to proceed with GCP deployments at least for now. We'll turn this toggle off when we migrate to public, sabre.com domains which have valid SSL certs